+++ dopplerview/pipeline/pipeline.py
@@ -110,8 +110,9 @@ class Context:
         
         logger.info(f"[Pipeline] Reading cache from {h5_cache_path}")
 
-        cache = h5_file.read_h5_to_dict(h5_cache_path)
+        cache, metadata = h5_file.read_h5_to_dict(h5_cache_path)
         self._init_cache(cache)
+        self.metadata.update(metadata)
 
     def ensure_directory(self, path):
         path = Path(path)
@@ -234,14 +235,23 @@ class Context:
         with self.lock:
             self.__cache[key] = (self.__cache[key][0], 'cached')
     
-    def export_cache(self, filepath):
-        cache = {}
-        with self.cache_lock:
-            for key, (value, status) in self.__cache.items():
-                if status == 'produced':  # Only export values that were produced and not yet cached
-                    cache[key] = value
-                    self.cache_value(key)  # Mark as cached after exporting
-        h5_file.write_dict_to_h5(cache, filepath, overwrite=False)
+    def cache_values(self, keys):
+        for key in keys:
+            if key in self.__cache and self.__cache[key][1] != 'cached':
+                self.cache_value(key)
+
+    def get_produced_values(self):
+        with self.lock:
+            return dict([(key, value) for key, (value, status) in self.__cache.items() if status == 'produced'])
+    
+    # def export_cache(self, filepath):
+    #     cache = {}
+    #     with self.cache_lock:
+    #         for key, (value, status) in self.__cache.items():
+    #             if status == 'produced':  # Only export values that were produced and not yet cached
+    #                 cache[key] = value
+    #                 self.cache_value(key)  # Mark as cached after exporting
+    #     h5_file.write_dict_to_h5(cache, filepath, overwrite=False)
 
 class Pipeline:
     def __init__(self, output_manager, debug_mode=False):