+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Move configuration in a specific window
 * Add visualization of to-be-processed inputs while highlighting the currently processed input, with a dedicated progress bar 
 * Add parameters for pre-mask method and and branch signal correction
+* Take config into account to validate/invalidate cache in debug mode
 
 ## 1.5.0
 * Pre-classify branch signals by clustering of their first harmonic feature in the complex domain